Add PAFF Smart Cut support for H.264 1080i50 streams
Full frame-accurate cutting support for H.264 PAFF (Picture-Adaptive
Frame-Field) encoded streams, as used by DVB broadcasters like kabel eins.

Key features:
- TTNaluParser: PAFF field-pair merging, independent PAFF detection
- TTESSmartCut: SPS Unification (rewrite encoder MBAFF output for source
  PAFF SPS), EOS DPB flush at re-encode→stream-copy transition, MMCO
  neutralization for first 32 stream-copy AUs (all slice types I/P/B)
- TTFFmpegWrapper: PAFF-aware frame indexing, field_pic_flag detection
- TTMkvMergeProvider: PAFF field-pair PTS assignment, per-packet detection
- ttcut-demux: correct frame rate for interlaced streams
- Smart Cut progress bar fix (was always 0%), elapsed time display
- Progress reporting standardized from permille to percent (0-100)
- Compiler warnings fixed (unused variables, deprecated Qt API, signedness)

Tested with Moon Crash (2022) PAFF and Clouseau MBAFF — 7/7 previews
decoder-error-free, no ad frame leaks, no regressions.

+++ avstream/ttfilebuffer.cpp
@@ -272,7 +272,7 @@ void TTFileBuffer::seekForward(quint64 offset)
 void TTFileBuffer::seekBackward(quint64 offset)
 {
   //printf("seek backward: %lld / %lld\n", offset, readPos);
-  readPos = (offset <= readPos) ? readPos - offset : 0;
+  readPos = (static_cast<qint64>(offset) <= readPos) ? readPos - static_cast<qint64>(offset) : 0;
 
   seekAbsolute(readPos);
 }
